re-organize tally report components in libs/ui
