+++ libs/ui/src/index.ts
@@ -9,8 +9,6 @@ export * from './button_list';
 export * from './card';
 export * from './centered_large_prose';
 export * from './contest_choice_button';
-export * from './contest_tally';
-export * from './contest_writein_tally';
 export * from './display_settings';
 export * from './error_boundary';
 export * from './election_info_bar';
@@ -42,8 +40,6 @@ export * from './main';
 export * from './mock_usb_drive';
 export * from './modal';
 export * from './number_pad';
-export * from './precinct_scanner_tally_report';
-export * from './precinct_scanner_tally_qrcode';
 export * from './printing_ballot_image';
 export * from './progress_bar';
 export * from './progress_ellipsis';
@@ -64,9 +60,7 @@ export * from './setup_card_reader_page';
 export * from './set_clock';
 export * from './tabbed_section';
 export * from './table';
-export * from './tally_report';
-export * from './tally_report_metadata';
-export * from './tally_report_summary';
+export * from './reports';
 export * from './test_mode';
 export * from './text';
 export * from './theme_manager_context';
@@ -81,9 +75,7 @@ export * from './unlock_machine_screen';
 export * from './system_administrator_screen_contents';
 export * from './unconfigure_machine_button';
 export * from './print_element';
-export * from './precinct_scanner_tally_reports';
 export * from './change_precinct_button';
-export * from './precinct_scanner_ballot_count_report';
 export * from './export_logs_modal';
 export * from './graphics';
 export * from './types';
