+++ libs/ui/src/reports/contest_writein_tally.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { ContestId, Election } from '@votingworks/types';
 import { electionSample2Fixtures } from '@votingworks/fixtures';
-import { render, screen, within } from '../test/react_testing_library';
+import { render, screen, within } from '../../test/react_testing_library';
 
 import { ContestWriteInTally } from './contest_writein_tally';
