+++ libs/ui/src/reports/precinct_scanner_tally_qrcode.test.tsx
@@ -13,7 +13,7 @@ import {
   compressTally,
 } from '@votingworks/utils';
 import { fakeKiosk, mockOf } from '@votingworks/test-utils';
-import { render, screen } from '../test/react_testing_library';
+import { render, screen } from '../../test/react_testing_library';
 
 import {
   getSignedQuickResultsReportingUrl,
