+++ libs/ui/src/reports/precinct_scanner_tally_report.test.tsx
@@ -16,7 +16,7 @@ import {
   filterTallyContestsByParty,
   singlePrecinctSelectionFor,
 } from '@votingworks/utils';
-import { render, screen, within } from '../test/react_testing_library';
+import { render, screen, within } from '../../test/react_testing_library';
 
 import { PrecinctScannerTallyReport } from './precinct_scanner_tally_report';
