+++ libs/ui/src/reports/precinct_scanner_tally_report.tsx
