+++ libs/ui/src/reports/precinct_scanner_tally_reports.test.tsx
@@ -11,7 +11,7 @@ import {
   getSubTalliesByPartyAndPrecinct,
   singlePrecinctSelectionFor,
 } from '@votingworks/utils';
-import { render, screen } from '../test/react_testing_library';
+import { render, screen } from '../../test/react_testing_library';
 import { PrecinctScannerTallyReports } from './precinct_scanner_tally_reports';
 
 test('polls closed: tally reports for each party in primary, single precinct', () => {
