+++ libs/ui/src/reports/tally_report.tsx
