+++ libs/ui/src/reports/tally_report_metadata.test.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import MockDate from 'mockdate';
 
 import { electionWithMsEitherNeither } from '@votingworks/fixtures';
-import { render } from '../test/react_testing_library';
+import { render } from '../../test/react_testing_library';
 
 import { TallyReportMetadata } from './tally_report_metadata';
