+++ libs/ui/src/reports/tally_report_metadata.tsx
@@ -4,7 +4,7 @@ import { Election } from '@votingworks/types';
 
 import { format } from '@votingworks/utils';
 
-import { Text } from './text';
+import { Text } from '../text';
 
 interface Props {
   election: Election;
